Xml support and new mappings

+++ src/main/java/org/example/task/cardealershop/service/CarServiceImpl.java
@@ -6,9 +6,12 @@ import org.example.task.cardealershop.entity.Client;
 import org.example.task.cardealershop.entity.Part;
 import org.example.task.cardealershop.exception.DuplicatedListEntityException;
 import org.example.task.cardealershop.exception.EntityByIdNotFoundException;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.util.ArrayList;
 import java.util.List;
 
 @Service
@@ -17,6 +20,7 @@ public class CarServiceImpl implements CarService {
     private CarRepository carRepository;
     private ClientService clientService;
     private PartService partService;
+    private Logger logger = LoggerFactory.getLogger(CarServiceImpl.class);
 
     @Autowired
     public CarServiceImpl(CarRepository carRepository, ClientService clientService, PartService partService) {
@@ -27,22 +31,34 @@ public class CarServiceImpl implements CarService {
 
     @Override
     public List<Car> getAllCars() {
+        logger.info("Getting list of all cars from database");
         return carRepository.findAll();
     }
 
     @Override
     public Car getCar(int carId) {
+        logger.info(String.format("Getting a car with id=%d from database", carId));
         return carRepository.findById(carId).orElseThrow(() -> new EntityByIdNotFoundException("Car", carId));
     }
 
     @Override
     public Car createCar(Car car) {
+        logger.info("Creating a new car");
         car.setId(0);
+        car.setClientList(getUpdatedClientList(car));
         return carRepository.save(car);
     }
 
+    private List<Client> getUpdatedClientList(Car car) {
+        List<Client> clients = new ArrayList<>();
+        car.getClientList().forEach((client -> clients.add(clientService.getClient(client.getId()))));
+        return clients;
+    }
+
+
     @Override
     public Car updateCar(Car updatedCar, int carId) {
+        logger.info(String.format("Updating info for a car with id=%d", carId));
         return carRepository.findById(carId)
                 .map((car) -> {
                     car.setModelName(updatedCar.getModelName());
@@ -50,6 +66,7 @@ public class CarServiceImpl implements CarService {
                     car.setEnginePower(updatedCar.getEnginePower());
                     car.setEnginePower(updatedCar.getEnginePower());
                     car.setColour(updatedCar.getColour());
+                    car.setClientList(getUpdatedClientList(updatedCar));
                     return carRepository.save(car);
                 })
                 .orElseThrow(() -> new EntityByIdNotFoundException("Car", carId));
@@ -57,16 +74,20 @@ public class CarServiceImpl implements CarService {
 
     @Override
     public void deleteCar(int carId) {
+        logger.info(String.format("Deleting a car with id=%d", carId));
+        if (!carRepository.existsById(carId)) throw new EntityByIdNotFoundException("Car", carId);
         carRepository.deleteById(carId);
     }
 
     @Override
     public List<Client> getClients(int carId) {
+        logger.info(String.format("Getting a list of clients for a car with id=%d", carId));
         return getCar(carId).getClientList();
     }
 
     @Override
     public List<Client> addClient(int carId, int clientId) {
+        logger.info(String.format("Adding client to a car with id=%d", carId));
         Car car = getCar(carId);
         List<Client> clientList = car.getClientList();
         Client client = clientService.getClient(clientId);
@@ -78,6 +99,7 @@ public class CarServiceImpl implements CarService {
 
     @Override
     public List<Client> deleteClient(int carId, int clientId) {
+        logger.info(String.format("Removing client from a car with id=%d", carId));
         Car car = getCar(carId);
         List<Client> clientList = car.getClientList();
         Client client = clientService.getClient(clientId);
@@ -89,11 +111,13 @@ public class CarServiceImpl implements CarService {
 
     @Override
     public List<Part> getParts(int carId) {
+        logger.info(String.format("Getting a list of parts for a car with id=%d", carId));
         return getCar(carId).getPartList();
     }
 
     @Override
     public List<Part> addPart(int carId, int partId) {
+        logger.info(String.format("Adding part to a car with id=%d", carId));
         Car car = getCar(carId);
         List<Part> partList = car.getPartList();
         Part part = partService.getPart(partId);
@@ -105,6 +129,7 @@ public class CarServiceImpl implements CarService {
 
     @Override
     public List<Part> deletePart(int carId, int partId) {
+        logger.info(String.format("Removing part from a car with id=%d", carId));
         Car car = getCar(carId);
         List<Part> partList = car.getPartList();
         Part part = partService.getPart(partId);
